+++ tools/sygus/CMakeLists.txt
@@ -1,12 +1,12 @@
 add_executable(sygustestparser testparser.cpp)
-target_link_libraries(sygustestparser ${Z3_LIBRARY} ${Boost_SYSTEM_LIBRARY} ${GMPXX_LIB} ${GMP_LIB})
+target_link_libraries(sygustestparser aevalSolver ${Z3_LIBRARY} ${Boost_SYSTEM_LIBRARY} ${GMPXX_LIB} ${GMP_LIB})
 #if (BISON_FOUND)
 add_dependencies(sygustestparser SyGuSParserBison)
 #endif()
 install(TARGETS sygustestparser RUNTIME DESTINATION bin)
 
 add_executable(sygussolver sygussolver.cpp)
-target_link_libraries(sygussolver ${Z3_LIBRARY} ${Boost_SYSTEM_LIBRARY} ${GMPXX_LIB} ${GMP_LIB})
+target_link_libraries(sygussolver aevalSolver ${Z3_LIBRARY} ${Boost_SYSTEM_LIBRARY} ${GMPXX_LIB} ${GMP_LIB})
 #if (BISON_FOUND)
 add_dependencies(sygussolver SyGuSParserBison)
 #endif()
